Refactor generate mannequin command
